runtime: share panic rethrow logic with wasm GC
